feat(eslint): sort destructuring keys
This is the last thing that is not sorted. By sorting this, we gain
consistency with all other "object literals". It also means that for
e.g. props in react components or "named arguments" for other functions,
the type definition and the destructuring of the arguments has the same
order, so it's easier to find the corresponding type definition